ml-service/app/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
pandas
scikit-learn
numpy
torch
psycopg2-binary
shap
langdetect
python-multipart
umap-learn
scipy
sqlalchemy
transformers
pydantic
accelerate
captum
lime
matplotlib
seaborn
bcrypt